+++ src/main/java/com/yuanno/block_clover/items/artifacts/MagicChangeArtifactItem.java
@@ -31,7 +31,7 @@ import com.yuanno.block_clover.spells.sealing.SealingProjectileAbility;
 import com.yuanno.block_clover.spells.slash.SlashBladesAbility;
 import com.yuanno.block_clover.spells.sword.AirDashAbility;
 import com.yuanno.block_clover.spells.time.TimeStealAbility;
-import com.yuanno.block_clover.spells.water.WaterBallAbility;
+import com.yuanno.block_clover.spells.water.waterball.WaterBallAbility;
 import com.yuanno.block_clover.spells.wind.WindBladeAbility;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
