+++ src/main/java/com/yuanno/block_clover/world/gen/ModEntityGeneration.java
@@ -1,5 +1,7 @@
 package com.yuanno.block_clover.world.gen;
 
+import com.yuanno.block_clover.api.BeJavapi;
+import com.yuanno.block_clover.api.Beapi;
 import com.yuanno.block_clover.init.ModEntities;
 import net.minecraft.entity.EntityClassification;
 import net.minecraft.util.RegistryKey;
@@ -18,7 +20,7 @@ public class ModEntityGeneration {
         RegistryKey key = RegistryKey.create(Registry.BIOME_REGISTRY, event.getName());
         Set types = BiomeDictionary.getTypes(key);
 
-        if (!types.contains(BiomeDictionary.Type.WET) && !types.contains(BiomeDictionary.Type.VOID)&& !types.contains(BiomeDictionary.Type.OCEAN) && !types.contains(BiomeDictionary.Type.RIVER)  && !types.contains(BiomeDictionary.Type.WATER))
+        if (!types.contains(BiomeDictionary.Type.WET) && !types.contains(BiomeDictionary.Type.VOID) && !types.contains(BiomeDictionary.Type.OCEAN) && !types.contains(BiomeDictionary.Type.RIVER)  && !types.contains(BiomeDictionary.Type.WATER))
         {
             event.getSpawns().addSpawn(EntityClassification.CREATURE, new MobSpawnInfo.Spawners(ModEntities.BANDIT.get(), 20, 2, 4));
         }
@@ -30,9 +32,9 @@ public class ModEntityGeneration {
         {
             event.getSpawns().addSpawn(EntityClassification.CREATURE, new MobSpawnInfo.Spawners(ModEntities.MONKEY_ENTITY.get(), 15, 2, 4));
         }
-        if (types.contains(BiomeDictionary.Type.OCEAN) || types.contains(BiomeDictionary.Type.BEACH))
+        if (types.contains(BiomeDictionary.Type.OCEAN))
         {
-            event.getSpawns().addSpawn(EntityClassification.CREATURE, new MobSpawnInfo.Spawners(ModEntities.CLOVER_SHARK.get(), 1, 1, 2));
+            event.getSpawns().addSpawn(EntityClassification.WATER_CREATURE, new MobSpawnInfo.Spawners(ModEntities.CLOVER_SHARK.get(), 1, 1, 2));
         }
     }
 }
